[IMP] shopinvader_api_cart: code traceability
When a schema is processed to get the value provided in the case of a partial update, it's better to always take the value from the properties of the schema instance. The use of the dictionary returned by the 'model_dump' method should be limited to determine if a field has been provided or not in the original json message. This approach allows a better traceability of the use of defined fields and ease code analysis by external tools.