[FIX] shopinvader_product_attribute_set: Ensure schemas are loaded at loading time
To ensure that the schemas classes are put into the registry of extendable models when the addon is loaded, the schema module must be imported at module root. Moreover, in the tests, since the schemas are extendable class, we must ensure that the schema registry is properly initialized and reset at end of the tests. When extending schemas, it's also a good practice to call the methods overriden in the module on the schema's base classes. This ensures that the override is taken into account by the registry